feat: Adds event calendar support (#259)
* build(deps): Upgrade yarn

* build(deps): Upgrade gems

* build(deps): Remove obsolete gems

* v0.0.0

* build(deps): Update semantic-release

* feat(events): add upcoming events page

made a page to to view upcoming events in the management tools

* feat(events): added page to make new events

* added gem.lock

* functionally complete calendar

calendar has all the features needed need to make it less janky and
possible add/remove features based on feedback

* feat(events):page to view and make upcoming events

created a page to view, make, and delete events. These events contain a
title, description, location, owner(s), if it is a public and/or all day
event and when it starts and stops.

* fix: corrected formating errors and removed all day option

made the code comply with houndci-bot requests and removed the all day
bool for a event

* fix: removed left over allDay refernce

* fix: corrected houndci format suggestion

* feat(event): added owners as a list of admins only

implemented owners in a similar way to message recipients it will only
allow for the selection of admin user as the owner(s) of a event. No
owners it also allowed.

* fix(events): refactored to work with latest 2.0

altered new manage sidebar to include events and fixed bug in event
model testing

* Fixed issue on desktop

* feat(event): removed unneeded form options

removed owner and public options from event creation form since they
are no longer relevent for the intended purpose of this schedule

* Map overflow scroll

Signed-off-by: Peter Kos <pkos91@icloud.com>

* Fix scroll issue

Signed-off-by: Peter Kos <pkos91@icloud.com>

* Fixed toggle button; fits screen, sidebar width

Sidebar width is better
Toggle button fills height of navbar

Signed-off-by: Peter Kos <pkos91@icloud.com>

* fix: Allows optional assets to be unset

* fix(events):added basic testing

added the basic tests and renamed "end" to "finish" in event

* Fixes test

* fix(event): tweak test in attempt to get working

* fix(event): got testing to work

for some reason there is a flow for how redirects works for a
controller. The index goes to the correct redirect but the create does
not.

* fix(event): fix hound errors and db password error

accidently pushed local db password and corrected hound errors

* fix(event): fix houndci error i missed

* feat: Redesigns OAuth2 portal (#404)

* feat: Redesigns OAuth2 portal

* fix: Some Hound issues

* Changes application verbiage

* Fixes ID issue

* Fixes table link and sign out button

* Peter's wishes

* Adds Redis dump files to .gitignore

* fix(event): fixed some of the pull request suggestions

* Shorten form :)

Signed-off-by: Peter Kos <pkos91@icloud.com>

* fix(forms): change size of all forms without textbox

changed the style of all forms that don't have a large text area or
other aspects of the form that make a full length form perferable. With
the exception of those forms all forms are now more narrow

* fix(event): implemented pull request changes

fixed cancel button and removed blue dot in list. decided to use simple
form datetime picker instead of html datetime-local input because
firefox does not support it.

* fix: Fixes MyMLH branding in config (#412)

* feat(settings): Simplifies Config (#409)

* feat(settings): Simplifies Config

* Resolves Hound issues

* Update app/views/manage/configs/index.html.haml

Co-authored-by: Peter Kos <pkos91@icloud.com>

* design: Increases the font-weight and reorganizes the page

* fix: Hound issues

Co-authored-by: Peter Kos <pkos91@icloud.com>

* feat: Adds support for digital-only hackathons (#395)

* feat: Adds support for digital-only hackathons

* fix: Provides clarity for what is removed from digital hackathons

Co-authored-by: Peter Kos <pkos91@icloud.com>

* event(fix): renamed events to schedule and reverted forms

change event to schedule on the mangage sidebar and undid change to
forms

* feat: Adds GitHub Action for releases (#426)

* feat: Adds GitHub Action CI testing (#423)

* feat: Adds GitHub Action CI testing

* fix: Adds sudo

* refactor: Changes naming

* Removes travis.yml

* docs: Adds ruby version comment

* fix(event): schedule now shows end time

do to issues with "end" being a keyword in rails and what fullcalendar
uses for events ending end time was previously not showing up

* fix(event): fix houndci complaints

* bundle install

* fixed styling errors and added check for finish and start time

* made finish time optional and made css fixes

* fix houndci complaint

* fix houndci complaint part 2

* removed finish time requirement in test

Co-authored-by: Chris Baudouin, Jr <chrisbaudouinjr@icloud.com>
Co-authored-by: Chris Baudouin, Jr <cjb5326@rit.edu>
Co-authored-by: Peter Kos <pkos91@icloud.com>